[rust] deserializing from a too-long buffer is more forgiving
Previously there would be a deserialization error if the buffer passed
to `Message::deser(buf: &[u8])` was longer than the size field plus
the 12 bytes for the LMCP header.

Now it instead is a check to make sure the buffer is _at least_ that
long, making it easier to deal with streaming situations where a
buffer might contain more than one LMCP message.