Support for FIXED_ARRAY_PRIMITIVE
The previous attempt to support FIXED_ARRAY_PRIMITIVE made illegal use of anonymous array types. This commit provides support for these array primitives.

Other FIXED_ARRAY_ types remain unimplemented. Rather than leaving the generation of illegal code, I’ve commented out the offending code and am throwing IllegalArgumentException. This will mean that the use of unsupported / unimplemented features will be found during generation, rather than during compilation. This will also draw attention to the places where changes need to be made.

Also fixed in this commit is an error in calculate packed size where there was an unmatched close paren. On inspection, this was probably meant to be matched with a cast to UInt32, so that was added.