Correct specs for factories (templates), so that calculateChecksum has the param specifying the last byte. Bodies of the factory packages now use that additional param.
Also, calculateChecksum no longer makes a physical copy of the buffer's byte array.
Instead, we call a new ByteBuffer function, named Checksum, that computes the modular checksum over the
slice of the byte array specified as parameters.

The routines for GetObjectSize and Validate are cleaners, now using a new ByteBuffer routines that gets a
UInt32 from a specific position specified.

In ByteBuffers, we clean up some comments and remove the previous support routines used for calculating
the checksum and getting the message object size. We add the Checksum function and the Get_UInt32 from a
specified position.