Fix Error in Package Generation
This error arises, nondeterministically, depending on the order in which MDMs are added to the list of MDMs to process. Java’s `File.listContents` method does not promise a particular order in which the contents of a directory will be returned. On many machines, the order that was chosen masked this error: CMASI’s sub-packaged (vehicle, impact, perceive) would be generated before CMASI was generated. In that case, the code worked correctly.

But when CMASI was generated first — with the correct contents of afrl-cmasi.ads — the later generation of vehicle, impact and perceive would overwrite afrl-cmasi.ads, because of the unconditional write. Making the write conditional on the file not existing solves the problem.