Rogers/casing fix (#6)
* fix casing for generated/copied files, for TB04-026

* Add docstrings, remove redundant calls.

Docstrings are added to elements that are introduced in this change set
or are updated by this change set. The intent of the docstrings is to
explain the purpose of the various elements and, where appropriate,
suggest how they might be improved going forward.

There were several instances of redundant calls to the `SPECIAL_CASING`
map or to `getDeconflictedName`. These have been eliminated. Similarly,
redundant calls to `toLowerCase` have been eliminated where possible.

The `descending_namespace_spec` method has been significantly edited to
improve code quality.

* Previously flagged redundancies are removed. Most variables have been
  marked final. There were many instances of reuse of local variable
  for a different purpose, which challenges readability and maintenance;
  these have been eliminated and protected against through the
  systematic use of `final`.

* Inefficient string concatenation (concatenating strings with
  variables) are replaced by use of `MessageFormat`.

* Multiline strings are (efficiently) broken across over multiple lines,
  for readability.

More could be done to improve `descending_namespace_spec`, specifically
to eliminate the for loops; possibly a recursive implementation would
fit well. That was not undertaken for this commit.

For testing, Ada sources were generated from all MDMs and spot-checked
for reasonableness.

Co-authored-by: rogers <rogers@adacore.com>