Correct a design flaw in the Raw_Bytes function returning a String value, in which the body truncated the result at Max_String_Length if the buffer content was larger than that. That result works correctly with the specific OpenUxAS application code because the app never exceeds strings of that maximum length, but in general it could omit existing buffer content. The function's resulting String value is unconstrained and is meant to contain the entire buffer content. Hence the function now returns the entire content, with a new precondition verifying that the complete content is not physically larger than the maximum possible Ada String (index) length.
Change the name of the function Raw_Bytes returning a String value (the one
described above) so that it is no longer overloaded with the function that returns
a Byte_Array value. The new function name is Raw_Bytes_As_String. Prior to the
name change, the only difference in the functions' profiles was the result type,
i.e., the names and parameters were identical. Calls to such functions are
disambiguated by the expected result type, but some call contexts in the
ByteBuffers package itself were ambiguous. In particular, the function returning
the Byte_Array value is used by the application code but also serves as the
"model" for the ByteBuffer type and therefore appears in some of the routines'
postconditions.

Correct a minor error in an output string in the functional test program
Test_ByteBuffers.

Create a new test main procedure used for proof of client code, in particular to
verify the completeness of the ByteBuffers routines' postconditions. This main
procedure, Prove_ByteBuffers, is SPARK compliant, unlike Test_ByteBuffers. Because
it is intended to prove only successful client use-cases, it contains a subset of
the checks in Test_ByteBufers.

Add the new main procedure Prove_ByteBuffers to the project GPR file, as the
default main. As the default, the SPARK-compliant Prove_ByteBuffers procedure will
be analyzed when we tell gnatprove to prove all the files.

Remove extraneous blanks in file avtas-lmcp-bytebuffers-copying.ads.